float en reto13 y reto09 sin mirar

+++ py6_retos2/reto13.py
@@ -1,5 +1,5 @@
-x = int(input('Escriba el primer numero al que desea hallarle la media: '))
-z = int(input('Escriba el segundo numero al que desea hallarle la media: '))
+x = float(input('Escriba el primer numero al que desea hallarle la media: '))
+z = float(input('Escriba el segundo numero al que desea hallarle la media: '))
 y = (x+z) / 2
 print('La media de',x,'y',z,'es igual a: ',y)
